+++ rust-sdk/client/src/gpa/position_bundle.rs
@@ -13,7 +13,7 @@ use std::error::Error;
 use solana_client::nonblocking::rpc_client::RpcClient;
 use solana_client::rpc_filter::Memcmp;
 use solana_client::rpc_filter::RpcFilterType;
-use solana_program::pubkey::Pubkey;
+use solana_pubkey::Pubkey;
 
 use crate::{generated::shared::DecodedAccount, PositionBundle};