+++ rust-sdk/client/src/pda/fusion_pools_config.rs
@@ -10,7 +10,7 @@
 
 use crate::generated::programs::FUSIONAMM_ID;
 use solana_program::program_error::ProgramError;
-use solana_program::pubkey::Pubkey;
+use solana_pubkey::Pubkey;
 
 pub fn get_fusion_pools_config_address() -> Result<(Pubkey, u8), ProgramError> {
     let seeds = &[b"config".as_ref()];