+++ rust-sdk/fusionamm-sdk/src/tests/anchor.rs
@@ -8,9 +8,8 @@
 // See the LICENSE file in the project root for license information.
 //
 
+use solana_pubkey::Pubkey;
 use std::{env::set_var, error::Error, fs::read_to_string, path::PathBuf, str::FromStr};
-
-use solana_sdk::pubkey::Pubkey;
 use toml::Table;
 
 pub fn anchor_programs(path: &str) -> Result<Vec<(String, Pubkey)>, Box<dyn Error>> {