+++ rust-sdk/fusionamm-sdk/src/tests/rpc.rs
@@ -16,6 +16,7 @@ use base64::engine::general_purpose::STANDARD;
 use base64::Engine;
 use fusionamm_client::{get_fusion_pools_config_address, FUSIONAMM_ID, FUSION_POOLS_CONFIG_DISCRIMINATOR};
 use serde_json::{from_value, to_value, Value};
+use solana_account::Account;
 use solana_account_decoder::{encode_ui_account, UiAccountEncoding};
 use solana_client::client_error::Result as ClientResult;
 use solana_client::{
@@ -26,21 +27,18 @@ use solana_client::{
     rpc_response::{Response, RpcBlockhash, RpcResponseContext, RpcVersionInfo},
     rpc_sender::{RpcSender, RpcTransportStats},
 };
+use solana_commitment_config::CommitmentLevel;
+use solana_epoch_info::EpochInfo;
+use solana_instruction::Instruction;
+use solana_keypair::Keypair;
+use solana_message::{v0::Message, VersionedMessage};
 use solana_program_test::tokio::sync::Mutex;
 use solana_program_test::{ProgramTest, ProgramTestContext};
-use solana_sdk::bs58;
-use solana_sdk::epoch_info::EpochInfo;
-use solana_sdk::{
-    account::Account,
-    commitment_config::CommitmentLevel,
-    instruction::Instruction,
-    message::{v0::Message, VersionedMessage},
-    pubkey::Pubkey,
-    signature::{Keypair, Signature},
-    signer::Signer,
-    system_program,
-    transaction::VersionedTransaction,
-};
+use solana_pubkey::Pubkey;
+use solana_sdk_ids::system_program;
+use solana_signature::Signature;
+use solana_signer::Signer;
+use solana_transaction::versioned::VersionedTransaction;
 use solana_version::Version;
 use spl_memo::build_memo;