+++ rust-sdk/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @crypticdot/fusionamm-rust-core
 
+## 1.0.50
+
 ## 1.0.49
 
 ### Patch Changes