+++ ts-sdk/client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @crypticdot/fusionamm-client
 
+## 1.0.50
+
 ## 1.0.49
 
 ## 1.0.48