+++ ts-sdk/integration/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crypticdot/fusionamm-integration
 
+## 1.0.50
+
+### Patch Changes
+
+- @crypticdot/fusionamm-core@1.0.50
+- @crypticdot/fusionamm-client@1.0.50
+- @crypticdot/fusionamm-sdk@1.0.50
+
 ## 1.0.49
 
 ### Patch Changes